+++ OpenNos.DAL.EF.MySQL/DB/OpenNos.edmx.sql
@@ -44,9 +44,9 @@
 -- -----------------------------------------------------------
 -- Entity Designer DDL Script for MySQL Server 4.1 and higher
 -- -----------------------------------------------------------
--- Date Created: 04/03/2016 20:24:47
+-- Date Created: 04/04/2016 00:55:31
 
--- Generated from EDMX file: C:\Users\Dominik\Source\Repos\OpenNos\OpenNos.DAL.EF.MySQL\DB\OpenNos.edmx
+-- Generated from EDMX file: C:\Users\ERWAN\Desktop\OpenNos Git\OpenNos.DAL.EF.MySQL\DB\OpenNos.edmx
 -- Target version: 3.0.0.0
 
 -- --------------------------------------------------
@@ -388,8 +388,8 @@ CREATE TABLE `npcmonster`(
 	`WaterResistance` tinyint NOT NULL, 
 	`LightResistance` tinyint NOT NULL, 
 	`DarkResistance` tinyint NOT NULL, 
-	`MaxHP` smallint NOT NULL, 
-	`MaxMP` smallint NOT NULL);
+	`MaxHP` int NOT NULL, 
+	`MaxMP` int NOT NULL);
 
 ALTER TABLE `npcmonster` ADD PRIMARY KEY (NpcMonsterVNum);
