+++ OpenNos.Data/NpcMonsterDTO.cs
@@ -41,8 +41,8 @@ namespace OpenNos.Data
         public sbyte WaterResistance { get; set; }
         public sbyte LightResistance { get; set; }
         public sbyte DarkResistance { get; set; }
-        public short MaxHP { get; set; }
-        public short MaxMP { get; set; }
+        public int MaxHP { get; set; }
+        public int MaxMP { get; set; }
 
         #endregion
     }
